Switch to isomorphism class
The idea is that there can be multiple ways of converting to text-builder. IOW, formats. We can catch that idea by providing specialized typeclasses implying formats. This however moves out of scope of this lib.

What we provide in this lib is supposed to be most neutral. Something that has only one way to be converted into the builder. I.e., unambiguously. This then implies that there should be a way to convert back. And thus we arrive at isomorphism.